chore: organize test files
